fix: `function` `lamdba` default environments
Closes #1171

Currently `lambda` and `function` are set and the user cannot disable those via system property or environment property (the latter is probably what they want to do as that it is easy to do in the AWS Lambda console). When you deploy to Lambda `lambda` `function` `ec2` and `cloud` are the active environments.

This PR considers  `lambda` and `function` default environments. Hence, if the user specifies for example `MICRONAUT_ENVIRONMENTS=prod` in the AWS console, the active environment for the lambda function will be only `prod`.

The only drawback is that right now in the test classpath `function`, `lambda` and `test` are the active environments with PR only `test` is the active enviroments. Default enviroments are not used if an environment is deduced.

I consider the current behaviour a bug and I think we should fix it even if we remove `lambda` and `function` for the tests.