fix: drive-by localhost cors deny TCK test (#1591)
When the filter blocks access to localhost from a non-local origin, there is no ROUTE_INFO set on the response object.

Previously, we just called get() on the Optional.  This fails.

I believe the fix is to only continue processing the response if the ROUTE_INFO has been applied to the response.