Add application/zip as a binary type and allow configuration (#1754)
Previously when we used serverless, we added application/zip to the list of binary types.

This was lost when we moved away from serverless, and was caught by a test in starter

https://github.com/micronaut-projects/micronaut-starter/pull/1853

This PR changes from a static set of binary types to a singleton bean that allows types to be
added to the set.  It also adds application/zip as a binary type by default.