fix: run Junit4 test into Junit5 test
This PR moves a Java test which was not being run as it was a JUnit 4 test to `src/test/java` and transforms it into a JUnit5 test.

It updates to latest Micronaut core as it was failing locally for me without this upgrade.

Removes the asciidoc tags as this file was not being embedded in the docs.